requirements.txt
google-api-python-client==2.90.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
oauth2client==4.1.3
Scrapy==2.9.0
scrapy-splash==0.9.0
